requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.64.0
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
requests>=2.28.0
lxml>=4.9.0
scikit-learn>=1.1.0
plotly>=5.15.0 